requirements.txt
flask==2.3.3
flask_sqlalchemy==3.0.5
flask_migrate==4.0.1
python-dotenv==1.0.0
psycopg2
marshmallow-sqlalchemy
Flask-JWT-Extended
